Started the page for bicycles and added function
Started the page for bicycles and the input boxes for the function and added function to calculate the burned calorie based on four factors: weight of the rider, average speed of the ride, the time of the ride and the intensity to decide the MET value which spans around 4-10 for bicycles and mine has 3 value 4, 6 and 8.  Nead to add a few sentences for each type of bike and add css for background and more readable or at least eye pleasuring view.